fix(rag): Address review findings on the database name validation
- Use re.fullmatch, not re.match: '$' also matches before a trailing newline,
  so "mydb\n" and a 64-character name ending in a newline passed the rule the
  entity advertises. Unreachable through the route today, since pydantic's rust
  regex engine anchors at end-of-haystack and rejects the path first, but the
  service and the tests call the classmethod directly.
- Move the new cases into their own test module. They exercise a pure predicate,
  yet the file they sat in has an autouse fixture that opens a Mongo connection,
  which would fail them at setup on a machine without one.
- Regenerate the SDK: the create route gained an OpenAPI description.
- Fix gender agreement in the French and Italian help strings.

Refs #1834
