+++ packages/core/swiss_ai_hub/core/persistence/rag/datalake/entities/tests/unit/test_bucket_entity_ingestor.py
@@ -112,17 +112,3 @@ class TestBucketNameValidation:
         bucket = BucketEntity.create_bucket(bucket_name="SharePointDocs")
 
         assert bucket.bucket_name == "SharePointDocs"
-
-
-class TestNewDatabaseNameValidation:
-    @pytest.mark.parametrize(
-        "name",
-        ["MyDb", "myDb", "1db", "ab", "a" * 64, "my-db", "my_db", "my db", "my.db", ""],
-    )
-    def test_rejects_a_name_no_store_would_accept(self, name):
-        with pytest.raises(ValidationError):
-            BucketEntity.validate_new_database_name(name)
-
-    @pytest.mark.parametrize("name", ["mydb", "abc", "db001", "a" * 63])
-    def test_accepts_a_lowercase_alphanumeric_name(self, name):
-        BucketEntity.validate_new_database_name(name)
